chore: dissect & extract injected/*/index.js
