+++ src/injected/content/notifications.js
@@ -3,20 +3,23 @@ import bridge from './bridge';
 
 const notifications = {};
 
-export function onNotificationCreate(options, realm) {
-  sendMessage({ cmd: 'Notification', data: options })
-  .then((nid) => { notifications[nid] = { id: options.id, realm }; });
-}
+bridge.addHandlers({
+  Notification(options, realm) {
+    sendMessage({ cmd: 'Notification', data: options })
+    .then((nid) => { notifications[nid] = { id: options.id, realm }; });
+  },
+});
 
-export function onNotificationClick(nid) {
-  const { id, realm } = notifications[nid] || {};
-  if (id) bridge.post({ cmd: 'NotificationClicked', data: id, realm });
-}
-
-export function onNotificationClose(nid) {
-  const { id, realm } = notifications[nid] || {};
-  if (id) {
-    bridge.post({ cmd: 'NotificationClosed', data: id, realm });
-    delete notifications[nid];
-  }
-}
+bridge.addBackgroundHandlers({
+  NotificationClick(nid) {
+    const { id, realm } = notifications[nid] || {};
+    if (id) bridge.post({ cmd: 'NotificationClicked', data: id, realm });
+  },
+  NotificationClose(nid) {
+    const { id, realm } = notifications[nid] || {};
+    if (id) {
+      bridge.post({ cmd: 'NotificationClosed', data: id, realm });
+      delete notifications[nid];
+    }
+  },
+});
