+++ src/injected/web/bridge.js
@@ -1,5 +1,14 @@
-import { noop } from '../utils/helpers';
+import { assign, noop } from '../utils/helpers';
+
+const handlers = {};
 
 export default {
   load: noop,
+  addHandlers(obj) {
+    assign(handlers, obj);
+  },
+  onHandle(obj) {
+    const handle = handlers[obj.cmd];
+    if (handle) handle(obj.data);
+  },
 };
