+++ src/injected/web/gm-api.js
@@ -1,0 +1,159 @@
+import { cache2blobUrl, getUniqId } from '#/common';
+import bridge from './bridge';
+import store from './store';
+import { onTabCreate } from './tabs';
+import { onRequestCreate } from './requests';
+import { onDownload } from './download';
+import { onNotificationCreate } from './notifications';
+import { decodeValue, dumpValue, loadValues } from './gm-values';
+import {
+  slice, objectKeys, atob, Error, jsonDump, log, utf8decode,
+} from '../utils/helpers';
+
+const { getElementById } = Document.prototype;
+const { lastIndexOf } = String.prototype;
+
+export function createGmApiProps() {
+  // these are bound to script data that we pass via |this|
+  const boundProps = {
+    GM_deleteValue(key) {
+      const { id } = this;
+      const values = loadValues(id);
+      dumpValue({
+        id, key, oldRaw: values[key],
+      });
+      delete values[key];
+    },
+    GM_getValue(key, def) {
+      const raw = loadValues(this.id)[key];
+      return raw ? decodeValue(raw) : def;
+    },
+    GM_listValues() {
+      return objectKeys(loadValues(this.id));
+    },
+    GM_setValue(key, val) {
+      const { id } = this;
+      const dumped = jsonDump(val);
+      const raw = dumped ? `o${dumped}` : null;
+      const values = loadValues(id);
+      dumpValue({
+        id, key, val, raw, oldRaw: values[key],
+      });
+      values[key] = raw;
+    },
+    GM_getResourceText(name) {
+      if (name in this.resources) {
+        const key = this.resources[name];
+        const raw = this.cache[this.pathMap[key] || key];
+        if (!raw) return;
+        const i = raw::lastIndexOf(',');
+        const lastPart = i < 0 ? raw : raw::slice(i + 1);
+        return utf8decode(atob(lastPart));
+      }
+    },
+    GM_getResourceURL(name) {
+      if (name in this.resources) {
+        const key = this.resources[name];
+        let blobUrl = this.urls[key];
+        if (!blobUrl) {
+          const raw = this.cache[this.pathMap[key] || key];
+          if (raw) {
+            blobUrl = cache2blobUrl(raw);
+            this.urls[key] = blobUrl;
+          } else {
+            blobUrl = key;
+          }
+        }
+        return blobUrl;
+      }
+    },
+    GM_log(...args) {
+      log('log', [this.script.meta.name || 'No name'], ...args);
+    },
+    GM_registerMenuCommand(cap, func) {
+      const { id } = this;
+      const key = `${id}:${cap}`;
+      store.commands[key] = func;
+      bridge.post({ cmd: 'RegisterMenu', data: [id, cap] });
+    },
+    GM_unregisterMenuCommand(cap) {
+      const { id } = this;
+      const key = `${id}:${cap}`;
+      delete store.commands[key];
+      bridge.post({ cmd: 'UnregisterMenu', data: [id, cap] });
+    },
+  };
+
+  const props = {
+    GM_addStyle(css) {
+      let el = false;
+      const callbackId = registerCallback((styleId) => {
+        el = document::getElementById(styleId);
+      });
+      bridge.post({ cmd: 'AddStyle', data: { css, callbackId } });
+      // Mock a Promise without the need for polyfill
+      // It's not actually necessary because DOM messaging is synchronous
+      // but we keep it for compatibility with VM's 2017-2019 behavior
+      // https://github.com/violentmonkey/violentmonkey/issues/217
+      el.then = callback => callback(el);
+      return el;
+    },
+    GM_openInTab(url, options) {
+      const data = options && typeof options === 'object' ? options : {
+        active: !options,
+      };
+      data.url = url;
+      return onTabCreate(data);
+    },
+    GM_xmlhttpRequest: onRequestCreate,
+    GM_download: onDownload,
+    GM_notification(text, title, image, onclick) {
+      const options = typeof text === 'object' ? text : {
+        text,
+        title,
+        image,
+        onclick,
+      };
+      if (!options.text) {
+        throw new Error('GM_notification: `text` is required!');
+      }
+      onNotificationCreate(options);
+    },
+    GM_setClipboard(data, type) {
+      bridge.post({
+        cmd: 'SetClipboard',
+        data: { type, data },
+      });
+    },
+  };
+  // convert to object property descriptors
+  [props, boundProps].forEach(target => {
+    objectKeys(target).forEach(k => {
+      target[k] = propertyFromValue(target[k]);
+    });
+  });
+  return { props, boundProps };
+}
+
+export function propertyFromValue(value) {
+  const prop = {
+    writable: false,
+    configurable: false,
+    value,
+  };
+  if (typeof value === 'function') value.toString = propertyToString;
+  return prop;
+}
+
+function propertyToString() {
+  return '[Violentmonkey property]';
+}
+
+function registerCallback(callback) {
+  const callbackId = getUniqId('VMcb');
+  store.callbacks[callbackId] = (payload) => {
+    callback(payload);
+    delete store.callbacks[callbackId];
+  };
+  return callbackId;
+}
