+++ src/injected/web/notifications.js
@@ -3,6 +3,24 @@ import bridge from './bridge';
 let lastId = 0;
 const notifications = {};
 
+bridge.addHandlers({
+  NotificationClicked(id) {
+    const options = notifications[id];
+    if (options) {
+      const { onclick } = options;
+      if (onclick) onclick();
+    }
+  },
+  NotificationClosed(id) {
+    const options = notifications[id];
+    if (options) {
+      delete notifications[id];
+      const { ondone } = options;
+      if (ondone) ondone();
+    }
+  },
+});
+
 export function onNotificationCreate(options) {
   lastId += 1;
   notifications[lastId] = options;
@@ -15,21 +33,4 @@ export function onNotificationCreate(options) {
       image: options.image,
     },
   });
-}
-
-export function onNotificationClicked(id) {
-  const options = notifications[id];
-  if (options) {
-    const { onclick } = options;
-    if (onclick) onclick();
-  }
-}
-
-export function onNotificationClosed(id) {
-  const options = notifications[id];
-  if (options) {
-    delete notifications[id];
-    const { ondone } = options;
-    if (ondone) ondone();
-  }
 }
