+++ src/injected/web/store.js
@@ -1,0 +1,6 @@
+export default {
+  commands: {},
+  values: {},
+  callbacks: {},
+  state: 0,
+};
